CAY-2665 Support for widespread SQL types that falls into Jdbc type OTHER  - Move value types to 'o.a.c.value' package
